Fix bugs with new code
So the tests work as I expect... sort of.  It seems that getting rusage times can cause things to be pretty slow, every now and then.  I am totally unsure of how to deal with that at all.  But otherwise, it seems as though my new code is working fine.