Swapped time value retrieval around to get truer times
At this point, I'd say the tool could be used to actually do benchmarks and retrieve resolution information reliably.  I should definitely add hardware resolution retrieval to the code (clock_getres(..) or whatever its calledm as well as rusage's counterpart, if it exists).  I should note that this tool's resolution retrieval tool may not be trusted for the rusage times, as they seem to sometimes spike, likely due to a process being rescheduled or whatever.  I guess its fair to say in general that benchmarking is a lot harder than it looks.